Change balance to ulong

+++ LifeSim/Data/PendingBet.cs
@@ -1,10 +1,10 @@
 ﻿namespace LifeSim.Data;
 
-public class PendingBet(string clientId, int amount, string betType, int initialCount, DateTime expiresAt)
+public class PendingBet(string clientId, ulong amount, string betType, int initialCount, DateTime expiresAt)
 {
     public readonly Guid Id = Guid.NewGuid();
     public readonly string ClientId = clientId;
-    public readonly int Amount = amount;
+    public readonly ulong Amount = amount;
     public readonly string BetType = betType;
     public readonly int InitialCount = initialCount;
     public readonly DateTime ExpiresAt = expiresAt;
